+++ plugin/pkg/auth/authorizer/rbac/bootstrappolicy/BUILD
@@ -43,13 +43,13 @@ go_test(
         "//pkg/apis/core/install:go_default_library",
         "//pkg/apis/rbac/install:go_default_library",
         "//pkg/apis/rbac/v1:go_default_library",
-        "//pkg/registry/rbac/validation:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/api/rbac/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/api/meta:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/diff:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/sets:go_default_library",
+        "//staging/src/k8s.io/component-helpers/auth/rbac/validation:go_default_library",
         "//vendor/sigs.k8s.io/yaml:go_default_library",
     ],
 )
