+++ staging/publishing/import-restrictions.yaml
@@ -256,5 +256,6 @@
   - k8s.io/api
   - k8s.io/apimachinery
   - k8s.io/client-go
+  - k8s.io/component-helpers
   - k8s.io/klog
   - k8s.io/utils
