+++ staging/src/k8s.io/component-helpers/auth/OWNERS
@@ -1,0 +1,8 @@
+# See the OWNERS docs at https://go.k8s.io/owners
+
+approvers:
+- sig-auth-api-approvers
+reviewers:
+- sig-auth-api-reviewers
+labels:
+- sig/auth
