+++ staging/src/k8s.io/component-helpers/auth/rbac/reconciliation/reconcile_role.go
@@ -25,7 +25,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/kubernetes/pkg/registry/rbac/validation"
+	"k8s.io/component-helpers/auth/rbac/validation"
 )
 
 type ReconcileOperation string
